pyproject.toml
[tool.poetry]
name = "o-complex-test-task"
version = "0.1.0"
description = ""
authors = ["Badmajor <v.balonkin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.1"
flask-socketio = "^5.3.6"
openmeteo-requests = "^1.2.0"
requests-cache = "^1.2.1"
retry-requests = "^2.0.0"
numpy = "^2.0.0"
pandas = "^2.2.2"
validators = "^0.33.0"
dataclasses-json = "^0.6.7"
pytest = "^8.2.2"
python-socketio = "^5.11.3"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.31"
flask-session = {extras = ["redis"], version = "^0.8.0"}
flask-restful = "^0.3.10"
eventlet = "^0.36.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
